+++ ParseSwift.playground/Pages/23 - Cloud Hook Triggers.xcplaygroundpage/Contents.swift
@@ -39,19 +39,6 @@ struct GameScore: ParseObject {
     }
 }
 
-//: It's recommended to place custom initializers in an extension
-//: to preserve the memberwise initializer.
-extension GameScore {
-
-    init(points: Int) {
-        self.points = points
-    }
-
-    init(objectId: String?) {
-        self.objectId = objectId
-    }
-}
-
 /*:
  Parse Hook Triggers can be created by conforming to
  `ParseHookFunctionable`.