+++ ParseSwift.playground/Pages/4 - User - Continued.xcplaygroundpage/Contents.swift
@@ -104,10 +104,6 @@ extension GameScore {
     init(points: Int) {
         self.points = points
     }
-
-    init(objectId: String?) {
-        self.objectId = objectId
-    }
 }
 
 //: Logging out - synchronously