+++ Tests/ParseSwiftTests/ParseLiveQueryTests.swift
@@ -28,10 +28,6 @@ class ParseLiveQueryTests: XCTestCase {
         init(points: Int) {
             self.points = points
         }
-
-        init(objectId: String?) {
-            self.objectId = objectId
-        }
     }
 
     class TestDelegate: ParseLiveQueryDelegate {